llms/model_capacity/requirements.txt
requests>=2.31
google-auth>=2.29
google-cloud-monitoring>=2.21
